+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Add StartupTime and Device.BootTime (#887) @lucas-zimerman
+- Link events to currently active span (#909) @Tyrrrz
 
 ### Changes